# Conflicts:
#	hack/test/kind/infra/build-operator.sh
#	lib.Makefile

+++ hack/test/kind/infra/build-operator.sh
@@ -3,26 +3,40 @@
 # This script produces a custom operator build that is used
 # in the ST and e2e tests in this repo.
 #
+# It caches the cloned operator repo across runs. Only the actual Go
+# compile and Docker image build run on subsequent invocations.
+#
 # When DEV_IMAGE_PATH and DEV_IMAGE_TAG are set, the operator is built
 # for pushing to a remote registry (no PullNever hack, custom tag/registry).
 # The operator will resolve images as:
 #   <DEV_IMAGE_REGISTRY>/<DEV_IMAGE_PATH>/<image>:<DEV_IMAGE_TAG>
 # e.g., docker.io/caseydavenport/node:my-feature
 
-# Clone the repository if needed.
 set -e
 REPO=${REPO:-tigera/operator}
 BRANCH=${BRANCH:-master}
 
-rm -rf operator/
-echo "Cloning https://github.com/${REPO} @ ${BRANCH}"
-git clone --depth=1 https://github.com/${REPO} -b ${BRANCH} operator
+# Reuse the existing clone if branch matches, otherwise start fresh.
+if [ -d operator/.git ]; then
+  existing_branch=$(git -C operator rev-parse --abbrev-ref HEAD 2>/dev/null || true)
+  if [ "$existing_branch" = "$BRANCH" ]; then
+    echo "Reusing cached operator clone (branch: ${BRANCH}), pulling latest..."
+    git -C operator fetch --depth=1 origin ${BRANCH}
+    git -C operator reset --hard origin/${BRANCH}
+  else
+    echo "Branch changed (${existing_branch} -> ${BRANCH}), re-cloning..."
+    rm -rf operator/
+    git clone --depth=1 https://github.com/${REPO} -b ${BRANCH} operator
+  fi
+else
+  echo "Cloning https://github.com/${REPO} @ ${BRANCH}"
+  git clone --depth=1 https://github.com/${REPO} -b ${BRANCH} operator
+fi
 
-# Modify the versions that are in-use to match our locally built images.
 pushd operator
 
 if [ -n "$COMMIT" ]; then
-  # If the latest operator has issues, fetch a known working commit as a workaround."
+  # If the latest operator has issues, fetch a known working commit as a workaround.
   echo "Fetch commit $COMMIT"
   git fetch origin $COMMIT
   git checkout $COMMIT
@@ -51,11 +65,11 @@ else
   build/_output/bin/gen-versions -os-versions=../calico_versions.yml > pkg/components/calico.go
   find . -name '*.go' | xargs sed -i 's/PullIfNotPresent/PullNever/g'
 
+  # Build an operator image for us to use and tag it with a local-only tag.
   make image
   docker tag tigera/operator:latest docker.io/tigera/operator:test-build
 fi
 
-# Clean up after ourselves.
 popd
-rm -rf operator/
+# Don't rm -rf operator/ -- keep it cached for next run.
 set +e
